Organize src/putit's  structure
